+++ source/index.test.ts
@@ -1,3 +1,4 @@
+import { CommonModule } from "@angular/common";
 import { Component } from "@angular/core";
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { By } from "@angular/platform-browser";
@@ -8,6 +9,7 @@ import { LuxonModule } from "./";
   template: `
     <span id="iso-to-iso">{{ '2006-01-02T15:04:05-07:00' | dateTimeFromIso | dateTimeToUtc | dateTimeToIso }}</span>
     <span id="ymd-to-dmy">{{ '2006-01-02' | dateTimeFromFormat:'yyyy-MM-dd' | dateTimeToFormat:'dd-MM-yyyy' }}</span>
+    <span id="timestamp-to-short">{{ 1136239445000 | dateTimeFromJsDate | dateTimeToJsDate | date:'mmss' }}</span>
   `
 })
 export class TestComponent {
@@ -25,6 +27,7 @@ describe("test component", () => {
         TestComponent
       ],
       imports: [
+        CommonModule,
         LuxonModule
       ]
     });
@@ -45,6 +48,13 @@ describe("test component", () => {
     fixture.detectChanges();
 
     expect(element.textContent).toBe("02-01-2006");
+  });
+
+  it("demonstrates timestamp -> Date -> DateTime -> Date", () => {
+    const element = fixture.debugElement.query(By.css("#timestamp-to-short")).nativeElement;
+    fixture.detectChanges();
+
+    expect(element.textContent).toBe("0405");
   });
 
 });