+++ source/pipes/date-time-from-js-date.test.ts
@@ -1,0 +1,31 @@
+import { DateTime } from "luxon";
+
+import { DateTimeFromJsDatePipe } from "./date-time-from-js-date";
+
+describe("DateTimeFromJsDatePipe", () => {
+  
+  let pipe: DateTimeFromJsDatePipe;
+
+  beforeEach(() => {
+    pipe = new DateTimeFromJsDatePipe();
+  });
+
+  describe("#transform", () => {
+    
+    it("transforms a JavaScript Date into a DateTime", () => {
+      const result = pipe.transform(new Date(1136239445000));
+
+      expect(result).toBeInstanceOf(DateTime);
+      expect(result.isValid).toBe(true);
+    });
+
+    it("transforms a timestamp (epoch milliseconds) into a DateTime", () => {
+      const result = pipe.transform(1136239445000);
+
+      expect(result).toBeInstanceOf(DateTime);
+      expect(result.isValid).toBe(true);
+    });
+
+  });
+
+});