+++ source/pipes/date-time-from-js-date.ts
@@ -1,0 +1,13 @@
+import { Pipe, PipeTransform } from "@angular/core";
+import { DateTime } from "luxon";
+
+@Pipe({
+  name: "dateTimeFromJsDate"
+})
+export class DateTimeFromJsDatePipe implements PipeTransform {
+
+  transform(value: any): DateTime {
+    return DateTime.fromJSDate(value);
+  }
+
+};