+++ source/pipes/date-time-to-js-date.ts
@@ -1,0 +1,13 @@
+import { Pipe, PipeTransform } from "@angular/core";
+import { DateTime } from "luxon";
+
+@Pipe({
+  name: "dateTimeToJsDate"
+})
+export class DateTimeToJsDatePipe implements PipeTransform {
+
+  transform(value: DateTime): Date {
+    return value.toJSDate();
+  }
+
+};