+++ typings/luxon.d.ts
@@ -1,12 +1,13 @@
 declare module "luxon" {
 
   export class DateTime {
-    
+
     static fromHTTP(text: string, options?: Object): DateTime;
+    static fromJSDate(date: Date | any, options?: Object): DateTime;
     static fromISO(text: string, options?: Object): DateTime;
     static fromRFC2822(text: string, options?: Object): DateTime;
     static fromString(text: string, format: string, options?: Object): DateTime;
-    
+
     readonly invalidReason: string | null;
     readonly isValid: boolean;
     readonly zoneName: string;
@@ -16,6 +17,7 @@ declare module "luxon" {
     toISODate(): string;
     toISOTime(options?: Object): string;
     toISOWeekDate(): string;
+    toJSDate(): Date;
     toLocal(): DateTime;
     toUTC(offset?: number, options?: Object): DateTime;