+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project are documented here. The format is based on
 [Keep a Changelog][keep-a-changelog].
 
+## Unreleased
+
+-   Added JavaScript Date <-> DateTime pipes.
+
 ## 0.0.1 (2017-10-13)
 
 -   Aligned to Luxon 0.0.4.